pyproject.toml
[project]
name = "proctoring"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ensemble-boxes>=1.0.9",
    "ipykernel>=6.29.5",
    "matplotlib>=3.10.1",
    "mediapipe>=0.10.21",
    "mtcnn>=1.0.0",
    "numpy>=1.26.4",
    "opencv-python>=4.11.0.86",
    "opencv-python-headless>=4.11.0.86",
    "pandas>=2.2.3",
    "pillow>=11.2.1",
    "scikit-learn>=1.6.1",
    "scipy>=1.15.2",
    "tensorflow>=2.18.1",
    "torch>=2.7.0",
    "torchaudio>=2.7.0",
    "torchvision>=0.22.0",
    "ultralytics>=8.3.126",
]
